+++ application/shared/view/timeclock/employee/remove.php
@@ -37,16 +37,16 @@
                     </thead>
                     <tbody>
                         <?php
-                        for ($i=0; $i<count($this->sys->template->all_employees); $i++) {
+                        for ($i=0; $i<count($this->sys->template->employees); $i++) {
                             ?>
                             <tr>
-                            <td onclick="employeeTableClicked('view', '<?php timeclock_employee::writeout($i, 'id'); ?>')"><?php timeclock_employee::writeout($i, 'lastname'); ?>, <?php timeclock_employee::writeout($i, 'firstname'); ?></td>
-                            <td onclick="employeeTableClicked('view', '<?php timeclock_employee::writeout($i, 'id'); ?>')"><?php timeclock_employee::writeout($i, 'uid'); ?></td>
-                            <td onclick="employeeTableClicked('view', '<?php timeclock_employee::writeout($i, 'id'); ?>')"><?php timeclock_employee::writeout($i, 'username'); ?></td>
+                            <td onclick="employeeTableClicked('view', '{employees[<?php echo $i; ?>]['employee_id']}')">{employees[<?php echo $i; ?>]['employee_lastname']}, {employees[<?php echo $i; ?>]['employee_firstname']}</td>
+                            <td onclick="employeeTableClicked('view', '{employees[<?php echo $i; ?>]['employee_id']}')">{employees[<?php echo $i; ?>]['employee_uid']}</td>
+                            <td onclick="employeeTableClicked('view', '{employees[<?php echo $i; ?>]['employee_id']}')">{employees[<?php echo $i; ?>]['employee_username']}</td>
                                 <td>
                                     <ul class="icons">
-                                    <li class="ui-state-default ui-corner-all" title=".ui-icon-pencil"><span class="ui-icon ui-icon-pencil" onclick="employeeTableClicked('edit', '<?php timeclock_employee::writeout($i, 'id'); ?>')"></span></li>
-                                    <li class="ui-state-default ui-corner-all" title=".ui-icon-trash"><span class="ui-icon ui-icon-trash" onclick="employeeTableClicked('trash', '<?php timeclock_employee::writeout($i, 'id'); ?>')"></span></li>
+                                    <li class="ui-state-default ui-corner-all" title=".ui-icon-pencil"><span class="ui-icon ui-icon-pencil" onclick="employeeTableClicked('edit', '{employees[<?php echo $i; ?>]['employee_id']}')"></span></li>
+                                    <li class="ui-state-default ui-corner-all" title=".ui-icon-trash"><span class="ui-icon ui-icon-trash" onclick="employeeTableClicked('trash', '{employees[<?php echo $i; ?>]['employee_id']}')"></span></li>
                                     </ul>
                                 </td>
                             </tr>
@@ -70,11 +70,11 @@
 <div class="remove_employee_dialog">
     <div class="dialog_text">
         Are you sure you want to remove<br />
-        <span class="bold"><?php timeclock_employee::writeout($this->sys->template->employee_id, 'firstname', 'by_id'); ?> <?php timeclock_employee::writeout($this->sys->template->employee_id, 'lastname', 'by_id'); ?></span>?
+        <span class="bold">{employees_by_id[<?php echo $this->sys->template->employee_id; ?>]['employee_firstname']} {employees_by_id[<?php echo $this->sys->template->employee_id; ?>]['employee_lastname']}?
     </div>
 
     <form class="remove_employee_form" method="post" action="">
-        <input type="hidden" name="employee_id" value="<?php timeclock_employee::writeout($this->sys->template->employee_id, 'id', 'by_id'); ?>" />
+        <input type="hidden" name="employee_id" value="{employees_by_id[<?php echo $this->sys->template->employee_id; ?>]['employee_id']}" />
         <input type="hidden" name="remove_employee" value="remove" />
     </form>
 </div>